feat(partitions): combine symlink and partitions lambda
This is just to simplify and streamline automation.
