--- sam/app/symlink.yaml
@@ -1,73 +1,0 @@
-AWSTemplateFormatVersion: "2010-09-09"
-Transform: AWS::Serverless-2016-10-31
-Description: "wolfeidau: aws billing symlink lambda which maintains the hive structure for Athena"
-
-Parameters:
-  AppName:
-    Type: String
-    Description: Application Name used to store configuration.
-  Stage:
-    Type: String
-    Description: The stage where the stac is running in, e.g., dev, prod.
-    Default: dev
-  Branch:
-    Type: String
-    Description: The branch the stack was deployed.
-    Default: master
-  Commit:
-    Type: String
-  ReportBucketName:
-    Description: The name of the bucket you have configured to receive CUR files.
-    Type: String
-  CurPrefix:
-    Type: String
-    Description: The prefix for the CUR files in the bucket, this should exclude the leading '/'.
-
-Conditions:
-  IsDev: !Equals [!Ref Stage, "dev"]
-  IsProd: !Equals [!Ref Stage, "prod"]
-
-Globals:
-  Function:
-    Runtime: go1.x
-    Timeout: 30
-    MemorySize: 256
-    Environment:
-      Variables:
-        STAGE: !Ref Stage
-        BRANCH: !Ref Branch
-        COMMIT: !Ref Commit
-        AWS_ACCOUNT_ID: !Ref "AWS::AccountId"
-        RAW_EVENT_LOGGING: !If [IsDev, "true", "false"]
-
-Resources:
-  SymlinkFunctionLogGroup:
-    Type: AWS::Logs::LogGroup
-    Properties:
-      LogGroupName: !Sub "/aws/lambda/${SymlinkFunction}"
-      RetentionInDays: !If [IsProd, 365, 14]
-
-  SymlinkFunction:
-    Type: AWS::Serverless::Function
-    Properties:
-      CodeUri: ../../dist/handler.zip
-      Handler: symlink-curs-lambda
-      Policies:
-        - S3CrudPolicy:
-            BucketName: !Ref ReportBucketName      
-      Events:
-        S3EventRule:
-          Type: EventBridgeRule
-          Properties:
-            Pattern:
-              source:
-                - "aws.s3"
-              "detail-type":
-                - "Object Created"
-              detail:
-                bucket:
-                  name:
-                    - !Ref ReportBucketName
-                object:
-                  key:
-                    - prefix: !Ref CurPrefix
